--- benchmark_observer_llm/src/core/__init__.py
@@ -1,7 +1,0 @@
-import logging
-
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(levelname)s %(asctime)s %(message)s",
-    datefmt="%Y-%m-%dT%H:%M:%S",
-)
